tmux: use Git repository names for window names
`tmux` lets me set the window name to `#{b:pane_current_path}`, the
basename of the current path. However, if I'm using Git worktrees, I'll
often be in a directory named after a branch, like `main` or `master`.
This is uninformative.

Therefore, this script prints the basename of a Git remote URL, with a
trailing `.git` removed, so that windows are named after the
repositories they're in.

TODO:
- Include a relative path within the repository?
- Include a branch name, or the name of the worktree?