Use mDNS for Raspberry Pi
This fixes the weechat IRC bouncer setup with Tailscale's "Use Tailscale
subnets" setting on, because $employer's AWS CIDR conflicts with my
router's 10.0.0.0/8 thing. Might need some work to force various things
to IPv6 in the future, seems to work OK for now though.

Thanks to @lf- for pointing me to this. Even though I resent having to
do it.